Handlers/Yggdrasil_Core/requirements.txt
gunicorn
flask
requests
redis
pyyaml
sqlalchemy
python-dotenv
pymysql
cryptography